% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_microdata}
\alias{download_microdata}
\title{Descarga al disco una base de datos de la ENE}
\usage{
download_microdata(year, month, directory, vars = "all")
}
\arguments{
\item{year}{año en formato integer}

\item{month}{mes en formato character}

\item{directory}{directorio local character}

\item{vars}{vector de characters}
}
\value{
guarad en memoria un archivo RData
}
\description{
Descarga al disco una base de datos de la ENE
}
