% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_indicators.R
\name{get_employment}
\alias{get_employment}
\title{Calcula el número de ocupados}
\usage{
get_employment(data)
}
\arguments{
\item{data}{\code{dataframe} de la ENE}
}
\value{
número de ocupados
}
\description{
Calcula el número de ocupados
}
