% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_indicators.R
\name{get_employment_rate}
\alias{get_employment_rate}
\title{Calcula la tasa de ocupación}
\usage{
get_employment_rate(data)
}
\arguments{
\item{data}{\code{dataframe} de la ENE}
}
\value{
tasa de desempleo
}
\description{
Calcula la tasa de ocupación
}
