% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_indicators.R
\name{get_informal}
\alias{get_informal}
\title{Calcula el número de trabajadores informaels}
\usage{
get_informal(data)
}
\arguments{
\item{data}{\code{dataframe} de la ENE}
}
\value{
número de trabajadores informales
}
\description{
Calcula el número de trabajadores informaels
}
