% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_indicators.R
\name{get_labour_force}
\alias{get_labour_force}
\title{Calcula la fuerza de trabajo}
\usage{
get_labour_force(data)
}
\arguments{
\item{data}{\code{dataframe} de la ENE}
}
\value{
fuerza de trabajo
}
\description{
Calcula la fuerza de trabajo
}
