% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_microdata}
\alias{get_microdata}
\title{Carga en memoria una base de datos de la ENE}
\usage{
get_microdata(year, month, vars = "all")
}
\arguments{
\item{year}{año en formato integer}

\item{month}{mes en formato character}

\item{vars}{vector de characters}
}
\value{
un \code{}
}
\description{
Carga en memoria una base de datos de la ENE
}
