% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_indicators.R
\name{get_parcipation_rate}
\alias{get_parcipation_rate}
\title{Calcula la tasa de participación laboral}
\usage{
get_parcipation_rate(data)
}
\arguments{
\item{data}{\code{dataframe} de la ENE}
}
\value{
tasa de participación
}
\description{
Calcula la tasa de participación laboral
}
