% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_indicators.R
\name{get_unemployment}
\alias{get_unemployment}
\title{Calcula el número de desocupados}
\usage{
get_unemployment(data)
}
\arguments{
\item{data}{\code{dataframe} de la ENE}
}
\value{
número de desocupados
}
\description{
Calcula el número de desocupados
}
