% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_indicators.R
\name{get_unemployment_rate}
\alias{get_unemployment_rate}
\title{Calcula la tasa de desocupación}
\usage{
get_unemployment_rate(data)
}
\arguments{
\item{data}{\code{dataframe} de la ENE}
}
\value{
tasa de desocupación
}
\description{
Calcula la tasa de desocupación
}
