% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_employment_time}
\alias{plot_employment_time}
\title{Genera un gráfico de la desocupación para 2 o más meses}
\usage{
plot_employment_time(df_list, design = "on")
}
\arguments{
\item{df_list}{lista de \code{dataframe}}

\item{design}{character con dos opciones: on u off}
}
\value{
objeto ggplot
}
\description{
Genera un gráfico de la desocupación para 2 o más meses
}
