% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{read_data_internal}
\alias{read_data_internal}
\title{Lee archivo en formato csv}
\usage{
read_data_internal(y, m, f)
}
\arguments{
\item{m}{un entero.}

\item{f}{un archivo}
}
\value{
un \code{dataframe}.
}
\description{
Lee archivo en formato csv
}
