% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{search_quarter_internal}
\alias{search_quarter_internal}
\title{Generar mes en formato numérico a partir de un character}
\usage{
search_quarter_internal(month_str)
}
\arguments{
\item{month_str}{mes en formato character}
}
\value{
mes en formato integer.
}
\description{
Generar mes en formato numérico a partir de un character
}
